#222 component-ize modals and create storybooks
