+++ src/components/modals/index.ts
@@ -1,0 +1,4 @@
+import ConfirmDeleteModal from './ConfirmDeleteModal';
+import EditProfileModal from './EditProfileModal';
+
+export { ConfirmDeleteModal, EditProfileModal };
