+++ src/screens/DashboardScreen.tsx
@@ -1,307 +1,15 @@
 import { useEffect, useState } from 'react';
 import { useSessionContext, useProfileContext } from '../hooks';
-import {
-    deleteProfileById,
-    updateProfileUsername,
-    getProfileQueue,
-    setProfileAdultFlag,
-    setProfileCountry,
-    removeProfileArray,
-} from '../supabase/profiles';
+import { deleteProfileById, getProfileQueue, removeProfileArray } from '../supabase/profiles';
 import { Navigate, useNavigate } from 'react-router-dom';
-import {
-    Box,
-    Button,
-    FilledInput,
-    FormControl,
-    InputLabel,
-    Modal,
-    Typography,
-} from '@mui/material';
-import {
-    ArrowBackIosNew,
-    Close,
-    Delete,
-    Edit,
-    Language,
-    Logout,
-    NoAdultContent,
-    WarningSharp,
-} from '@mui/icons-material';
-import { Profile, Session, ShowData } from '../types';
-import { ShowCarousel } from '../components';
+import { Button, Typography } from '@mui/material';
+import { Delete, Logout } from '@mui/icons-material';
+import { ShowData } from '../types';
+import { ConfirmDeleteModal, EditProfileModal, ShowCarousel } from '../components';
 import { SUPABASE, getMovieDetails, getTvDetails } from '../helpers';
 import Logger from '../logger';
 
 const LOG = new Logger('DashboardScreen');
-
-/**
- * A modal that allows the user to edit details
- * about their profile including the username,
- * country of origin and adult flag status.
- *
- * This will render to the screen as a button that
- * will open the modal when clicked.
- * @returns {JSX.Element}
- */
-function EditProfileModal({
-    session,
-    profile,
-    setProfile,
-}: {
-    session: Session;
-    profile: Profile;
-    setProfile: React.Dispatch<React.SetStateAction<Profile | null>>;
-}): JSX.Element {
-    const [open, setOpen] = useState(false);
-    const [username, setUsername] = useState('');
-    const [usernameError, setUsernameError] = useState(false);
-    const [country, setCountry] = useState('');
-    const [countryError, setCountryError] = useState(false);
-    const [isAdult, setIsAdult] = useState<boolean>(session.user.adult || false);
-
-    const handleOpen = () => {
-        setOpen(true);
-    };
-
-    const handleClose = () => {
-        setUsername('');
-        setUsernameError(false);
-        setCountry('');
-        setCountryError(false);
-        setOpen(false);
-    };
-
-    const changeUsername = async () => {
-        if (session && username.length > 2) {
-            const data = await updateProfileUsername(session.user.id, username);
-            setProfile(data);
-            setUsername('');
-        } else {
-            setUsernameError(true);
-        }
-    };
-
-    const toggleAdultFlag = async () => {
-        if (session) {
-            const data = await setProfileAdultFlag(session.user.id, !isAdult);
-            setProfile(data);
-            setIsAdult(!isAdult);
-        }
-    };
-
-    const changeCountry = async () => {
-        if (session && country.length === 2) {
-            const data = await setProfileCountry(session.user.id, country);
-            setProfile(data);
-            setCountry('');
-        } else {
-            setCountryError(true);
-        }
-    };
-
-    return (
-        <>
-            <Button
-                variant='contained'
-                size='large'
-                type='button'
-                color='secondary'
-                startIcon={<Edit />}
-                sx={{ m: 0.5, width: 210 }}
-                onClick={handleOpen}
-            >
-                Edit Profile
-            </Button>
-            <Modal
-                open={open}
-                onClose={handleClose}
-                aria-labelledby='modal-modal-title'
-                aria-describedby='modal-modal-description'
-            >
-                <Box
-                    sx={{
-                        position: 'absolute',
-                        top: '50%',
-                        left: '50%',
-                        transform: 'translate(-50%, -50%)',
-                        borderRadius: 2,
-                        width: {
-                            xs: 250,
-                            sm: 500,
-                        },
-                        boxShadow: 24,
-                    }}
-                >
-                    <div className='flex flex-col items-center bg-background p-4 rounded-md'>
-                        <Typography variant='h5'>Edit Profile</Typography>
-                        <FormControl sx={{ m: 0.5 }} variant='filled'>
-                            <InputLabel htmlFor='username' color='secondary' className='!text-text'>
-                                Change Username
-                            </InputLabel>
-                            <FilledInput
-                                name='username'
-                                value={username}
-                                onChange={(e) => {
-                                    setUsername(e.target.value);
-                                }}
-                                error={usernameError}
-                                onFocus={() => setUsernameError(false)}
-                                inputProps={{ minLength: 3 }}
-                                sx={{ m: 0.5, width: 210 }}
-                            />
-                        </FormControl>
-                        <Typography>Current Username: {profile?.username}</Typography>
-                        <Button
-                            variant='contained'
-                            type='button'
-                            color='secondary'
-                            startIcon={<Edit />}
-                            sx={{ m: 0.5, width: 210, mb: 2 }}
-                            onClick={changeUsername}
-                        >
-                            Change Username
-                        </Button>
-                        <FormControl sx={{ m: 0.5 }} variant='filled'>
-                            <InputLabel htmlFor='country' color='secondary' className='!text-text'>
-                                Change Country
-                            </InputLabel>
-                            <FilledInput
-                                name='country'
-                                value={country}
-                                onChange={(e) => {
-                                    setCountry(e.target.value);
-                                }}
-                                error={countryError}
-                                onFocus={() => setCountryError(false)}
-                                inputProps={{ maxLength: 2, minLength: 2 }}
-                                sx={{ m: 0.5, width: 210 }}
-                            />
-                        </FormControl>
-                        <Typography>Current Country: {profile?.country}</Typography>
-                        <Button
-                            variant='contained'
-                            type='button'
-                            color='secondary'
-                            sx={{ m: 0.5, width: 210, mb: 2 }}
-                            onClick={changeCountry}
-                            startIcon={<Language />}
-                        >
-                            Change Country
-                        </Button>
-                        <Typography>
-                            Adult Content: {profile?.adult ? 'visible' : 'not visible'}
-                        </Typography>
-                        <Button
-                            variant='contained'
-                            type='button'
-                            color='secondary'
-                            startIcon={<NoAdultContent />}
-                            sx={{ m: 0.5, mb: 2 }}
-                            onClick={toggleAdultFlag}
-                        >
-                            Toggle Adult Flag
-                        </Button>
-                        <Button
-                            variant='contained'
-                            type='button'
-                            color='error'
-                            startIcon={<Close />}
-                            sx={{ m: 0.5 }}
-                            onClick={handleClose}
-                        >
-                            Close
-                        </Button>
-                    </div>
-                </Box>
-            </Modal>
-        </>
-    );
-}
-
-/**
- * A modal to confirm if the user wants to permanently
- * delete their account.
- *
- * @returns {JSX.Element}
- */
-const ConfirmDeleteModal = ({ deleteProfile }: { deleteProfile: () => void }): JSX.Element => {
-    const [open, setOpen] = useState(false);
-
-    const handleOpen = () => {
-        setOpen(true);
-    };
-
-    const handleClose = () => {
-        setOpen(false);
-    };
-
-    return (
-        <>
-            <Button
-                variant='contained'
-                size='large'
-                color='error'
-                type='button'
-                sx={{ m: 0.5, width: 210 }}
-                startIcon={<WarningSharp />}
-                onClick={handleOpen}
-            >
-                Delete Profile
-            </Button>
-            <Modal
-                open={open}
-                onClose={handleClose}
-                aria-labelledby='modal-modal-title'
-                aria-describedby='modal-modal-description'
-            >
-                <Box
-                    sx={{
-                        position: 'absolute',
-                        top: '50%',
-                        left: '50%',
-                        transform: 'translate(-50%, -50%)',
-                        borderRadius: 2,
-                        width: {
-                            xs: 250,
-                            sm: 500,
-                        },
-                        boxShadow: 24,
-                    }}
-                >
-                    <div className='flex flex-col items-center bg-background p-4 rounded-md'>
-                        <Typography variant='h6' align='center'>
-                            Are you sure you want to delete your profile?
-                        </Typography>
-                        <Typography mb={2}>⚠️ Warning! This action cannot be undone.</Typography>
-                        <Button
-                            variant='contained'
-                            size='large'
-                            type='button'
-                            color='error'
-                            startIcon={<Delete />}
-                            sx={{ m: 0.5, width: 210 }}
-                            onClick={deleteProfile}
-                        >
-                            Yes
-                        </Button>
-                        <Button
-                            variant='contained'
-                            size='large'
-                            type='button'
-                            color='secondary'
-                            startIcon={<ArrowBackIosNew />}
-                            sx={{ m: 0.5, width: 210 }}
-                            onClick={handleClose}
-                        >
-                            No
-                        </Button>
-                    </div>
-                </Box>
-            </Modal>
-        </>
-    );
-};
 
 /**
  * A logged in users profile screen. This is used to display
@@ -317,6 +25,7 @@ export default function DashboardScreen(): JSX.Element {
     const navigate = useNavigate();
 
     LOG.debug(JSON.stringify(profile));
+    LOG.debug(JSON.stringify(session));
 
     const fallbackText = 'Your queue is empty! Add shows to your watch queue to view them here.';
