+++ src/stories/constants.ts
@@ -1,6 +1,6 @@
-import { ShowData } from '../types';
+import { Profile, Session, ShowData } from '../types';
 
-export const PROFILE = {
+export const PROFILE: Profile = {
     id: '1234',
     username: 'TestUser',
     email: 'example@no.mail',
@@ -12,6 +12,48 @@ export const PROFILE = {
     country: 'US',
     updated_at: '2023-08-09T03:17:57.310334+00:00',
     created_at: '2023-02-15T03:59:59.075061+00:00',
+};
+
+export const SESSION: Session = {
+    access_token: 'token',
+    token_type: 'bearer',
+    expires_in: 3600,
+    expires_at: 1693865876,
+    refresh_token: 'token',
+    user: {
+        id: 'token',
+        aud: 'authenticated',
+        role: 'authenticated',
+        email: 'test@no.co',
+        email_confirmed_at: '2023-09-04T21:17:55.537437973Z',
+        phone: '',
+        last_sign_in_at: '2023-09-04T21:17:55.539816463Z',
+        app_metadata: {
+            provider: 'email',
+            providers: ['email'],
+        },
+        user_metadata: {
+            username: 'test',
+        },
+        identities: [
+            {
+                id: 'token',
+                user_id: 'token',
+                identity_data: {
+                    email: 'test@no.co',
+                    sub: 'token',
+                },
+                provider: 'email',
+                last_sign_in_at: '2023-09-04T21:17:55.534787863Z',
+                created_at: '2023-09-04T21:17:55.534822Z',
+                updated_at: '2023-09-04T21:17:55.534822Z',
+            },
+        ],
+        created_at: '2023-09-04T21:17:55.532453Z',
+        updated_at: '2023-09-04T21:17:55.541357Z',
+        country: 'US',
+        adult: false,
+    },
 };
 
 export const MOVIE_DATA: ShowData = {
