+++ src/types/supabase.ts
@@ -4,7 +4,7 @@
 export interface User {
     app_metadata: unknown;
     aud: string;
-    confirmation_sent_at: string;
+    confirmation_sent_at?: string;
     confirmed_at?: string;
     created_at: string;
     email: string;
@@ -17,6 +17,7 @@ export interface User {
     user_metadata: unknown;
     adult: boolean | null;
     country: string | null;
+    last_sign_in_at: string;
 }
 
 /**
